+++ CHANGELOG.md
@@ -2,4 +2,6 @@
 
 ## [Unreleased]
 
+- LayerContentMetadata values (build, cache, launch) are now under a "types" key
+
 ## [0.3.0] 2021/09/17